use local Transmission RPC Lite library
Tried using the system Transmission RPC library (Ubuntu, src from https://bitbucket.org/blueluna/transmissionrpc/wiki/Home) but it
has the nasty habbit of downloading torrent links locally and upload them as a data file to Transmission. Somehow this fails
resulting in an 'invalid or corrupt torrent file' error from Transmission. In the 'blueluna' source a comment is made that Transmission
itself has difficulty donwloading the files but I could not reproduce this.

Credits for the Transmission RPC Lite library go to the CouchPotatoServer project.